[Issue 38] move CLI stuff including supervisor to new cli app
