+++ elixir/apps/cli/.gitignore
@@ -1,6 +1,6 @@
-/cover/
-/doc/
-/.fetch
-erl_crash.dump
-*.ez
-cli-*.tar
+# ignore the escript
+/fractals
+
+# ignore the generated images
+/images
+/test/images
