--- elixir/apps/cli/lib/cli/application.ex
@@ -1,20 +1,0 @@
-defmodule CLI.Application do
-  # See https://hexdocs.pm/elixir/Application.html
-  # for more information on OTP Applications
-  @moduledoc false
-
-  use Application
-
-  def start(_type, _args) do
-    # List all child processes to be supervised
-    children = [
-      # Starts a worker by calling: CLI.Worker.start_link(arg)
-      # {CLI.Worker, arg},
-    ]
-
-    # See https://hexdocs.pm/elixir/Supervisor.html
-    # for other strategies and supported options
-    opts = [strategy: :one_for_one, name: CLI.Supervisor]
-    Supervisor.start_link(children, opts)
-  end
-end
