--- elixir/apps/cli/test/cli_test.exs
@@ -1,8 +1,0 @@
-defmodule CLITest do
-  use ExUnit.Case
-  doctest CLI
-
-  test "greets the world" do
-    assert CLI.hello() == :world
-  end
-end
