+++ elixir/apps/cli/test/inputs/integration_mandelbrot.yml
