+++ elixir/apps/fractals/.gitignore
@@ -1,6 +1,2 @@
-# ignore the escript
-/fractals
-
 # ignore images (.keep files manually added)
-/images
 /test/images
